Add tenant field automatically in tenant_app models that do not have tenant field
