+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.9.0
+current_version = 2.0.0
 commit = True
 tag = True
